refactor(app): split service endpoint handlers into separate files
